:white_check_mark: [#2474] move iot admin tests to documenten component
