+++ src/openzaak/components/catalogi/tests/admin/__init__.py
@@ -1,2 +1,4 @@
 # SPDX-License-Identifier: EUPL-1.2
 # Copyright (C) 2019 - 2020 Dimpact
+
+# Informatieobjecttype admin tests have been moved to src/openzaak/components/documenten/tests/admin/besluiten because the catalogi API has been deprecated and the resources have been moved
