+++ src/openzaak/components/catalogi/tests/admin/test_published_readonly.py
@@ -21,7 +21,6 @@ from openzaak.tests.utils.admin import AdminTestMixin
 
 from ..factories import (
     BesluitTypeFactory,
-    InformatieObjectTypeFactory,
     ResultaatTypeFactory,
     StatusTypeFactory,
     ZaakTypeFactory,
@@ -107,32 +106,6 @@ class ReadonlyAdminTests(
 
         self.assertEqual("datum_einde_geldigheid" in form_fields, True)
         for field in besluittype_fields:
-            self.assertEqual(field in form_fields, False)
-
-    def test_readonly_informatieobjecttype(self, m):
-        """
-        check that in case of published informatieobjecttype only "datum_einde_geldigheid" field is editable
-        """
-        mock_selectielijst_oas_get(m)
-
-        informatieobjecttype = InformatieObjectTypeFactory.create(concept=False)
-        url = reverse(
-            "admin:catalogi_informatieobjecttype_change",
-            args=(informatieobjecttype.pk,),
-        )
-
-        response = self.app.get(url)
-
-        form = response.forms["informatieobjecttype_form"]
-        form_fields = list(form.fields.keys())
-        informatieobjecttype_fields = [
-            f.name
-            for f in informatieobjecttype._meta.get_fields()
-            if f.name != "datum_einde_geldigheid"
-        ]
-
-        self.assertEqual("datum_einde_geldigheid" in form_fields, True)
-        for field in informatieobjecttype_fields:
             self.assertEqual(field in form_fields, False)
 
     def test_readonly_statustype(self, m):
