+++ src/openzaak/components/documenten/tests/admin/informatieobjecttypen/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: EUPL-1.2
+# Copyright (C) 2019 - 2020 Dimpact
+
+# Catalogi API has been deprecated and informatieobjecttype has been moved to the Documenten API
+# Because of the default serializer view_name 's the admin already uses document urls.
+# to make it more clear the informatieobjecttype admin tests have already been moved to documenten admin tests.
