+++ src/openzaak/components/documenten/tests/admin/informatieobjecttypen/test_admin_publish.py
@@ -1,0 +1,112 @@
+# SPDX-License-Identifier: EUPL-1.2
+# Copyright (C) 2019 - 2020 Dimpact
+
+from django.contrib.auth.models import Permission
+from django.test import tag
+from django.urls import reverse
+from django.utils.translation import gettext as _
+
+import requests_mock
+from django_webtest import WebTest
+from maykin_2fa.test import disable_admin_mfa
+
+from openzaak.accounts.tests.factories import SuperUserFactory, UserFactory
+from openzaak.components.catalogi.tests.factories import (
+    InformatieObjectTypeFactory,
+)
+from openzaak.notifications.tests.mixins import NotificationsConfigMixin
+from openzaak.selectielijst.models import ReferentieLijstConfig
+from openzaak.selectielijst.tests.mixins import ReferentieLijstServiceMixin
+from openzaak.tests.utils import ClearCachesMixin
+
+
+@disable_admin_mfa()
+@requests_mock.Mocker()
+class ZaaktypeAdminTests(
+    NotificationsConfigMixin, ReferentieLijstServiceMixin, ClearCachesMixin, WebTest
+):
+    @classmethod
+    def setUpTestData(cls):
+        super().setUpTestData()
+
+        # there are TransactionTestCases that truncate the DB, so we need to ensure
+        # there are available years
+        config = ReferentieLijstConfig.get_solo()
+        config.allowed_years = [2017, 2020]
+        config.save()
+
+        cls.user = SuperUserFactory.create()
+
+    def setUp(self):
+        super().setUp()
+
+        self.app.set_user(self.user)
+
+    def test_publish_informatieobjecttype(self, m):
+        iot = InformatieObjectTypeFactory.create(
+            concept=True, vertrouwelijkheidaanduiding="openbaar"
+        )
+        iot.zaaktypeinformatieobjecttype_set.all().delete()
+        url = reverse("admin:catalogi_informatieobjecttype_change", args=(iot.pk,))
+
+        response = self.app.get(url)
+
+        # Verify that the publish button is visible and enabled
+        publish_button = response.html.find("input", {"name": "_publish"})
+        self.assertIsNotNone(publish_button)
+        publish_button = response.html.find(
+            "input", {"name": "_publish", "disabled": "disabled"}
+        )
+        self.assertIsNone(publish_button)
+
+        form = response.forms["informatieobjecttype_form"]
+
+        response = form.submit("_publish").follow()
+
+        iot.refresh_from_db()
+        self.assertFalse(iot.concept)
+
+        # Verify that the publish button is disabled
+        publish_button = response.html.find(
+            "input", {"name": "_publish", "disabled": "disabled"}
+        )
+        self.assertIsNotNone(publish_button)
+
+
+@tag("readonly-user")
+@disable_admin_mfa()
+class ReadOnlyUserTests(ClearCachesMixin, WebTest):
+    @classmethod
+    def setUpTestData(cls):
+        super().setUpTestData()
+
+        user = UserFactory.create(is_staff=True)
+        view_zaaktype = Permission.objects.get(codename="view_zaaktype")
+        view_informatieobjecttype = Permission.objects.get(
+            codename="view_informatieobjecttype"
+        )
+        view_besluittype = Permission.objects.get(codename="view_besluittype")
+        user.user_permissions.add(
+            view_zaaktype, view_informatieobjecttype, view_besluittype
+        )
+
+        cls.user = user
+
+    def setUp(self):
+        super().setUp()
+        self.app.set_user(self.user)
+
+    def test_informatieobjecttype_publish_not_possible(self):
+        informatieobjecttype = InformatieObjectTypeFactory.create(concept=True)
+        url = reverse(
+            "admin:catalogi_informatieobjecttype_change",
+            args=(informatieobjecttype.pk,),
+        )
+
+        detail_page = self.app.get(url)
+        form = detail_page.forms["informatieobjecttype_form"]
+
+        self.assertNotIn(_("Publiceren"), form.html)
+
+        # try to submit it anyway
+        form.submit("_publish", status=403)
